♻️ (dashboard): Move the step-2 probe transcript builder into its own module
buildProbeLines was exported from the component file, which trips
react-refresh/only-export-components — a warning, and `pnpm lint` runs
with --max-warnings 0.

Its own module suits it better anyway: the classification of a probe
outcome into ok / warn / err lines is the part worth testing directly,
and it now has a test file that exercises every branch without rendering
the step.

Refs AAASM-5132
